requirements.txt
asgiref==3.3.4
Django==3.2.4


# config
django-environ
django_extensions

# drf
djangorestframework
pytz==2021.1
sqlparse==0.4.1
werkzeug
